# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Tadweerah API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /notifications:
    get:
      operationId: getNotifications
      tags: [notifications]
      summary: Get company notifications (newest first, max 100)
      parameters:
        - in: query
          name: unread
          schema:
            type: boolean
          description: If true, only returns unread notifications.
      responses:
        "200":
          description: List of notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
        "401":
          description: Unauthorized
  /notifications/read-all:
    post:
      operationId: markAllNotificationsRead
      tags: [notifications]
      summary: Mark all company notifications as read
      responses:
        "200":
          description: All notifications marked read
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
        "401":
          description: Unauthorized
  /notifications/{notification_id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a single notification as read
      parameters:
        - in: path
          name: notification_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Notification marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
        "401":
          description: Unauthorized
        "404":
          description: Notification not found

  # ── Admin lookup write endpoints ─────────────────────────────────────────────
  # Protected by X-Admin-Key header. Key must match ADMIN_API_KEY env var.
...
components:
  schemas:
...
    Notification:
      type: object
      description: >
        In-app notification for a company user.
        Created automatically by the platform on key events.
      properties:
        id:
          type: string
        company_id:
          type: string
        type:
          type: string
          description: Stable internal type key (e.g. offer_received, outbid, offer_accepted).
        title_ar:
          type: string
        title_en:
          type: string
        body_ar:
          type: string
        body_en:
          type: string
        is_read:
          type: boolean
        related_entity_type:
          type: string
          description: "Entity type the notification links to (e.g. listing, deal)"
        related_entity_id:
          type: string
          description: UUID of the related entity
        created_at:
          type: string
          format: date-time
        read_at:
          type: string
          format: date-time
      required: [id, company_id, type, title_ar, title_en, is_read, created_at]