# lib/api-zod/src/generated/types/notification.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tadweerah API
 * OpenAPI spec version: 0.1.0
 */

/**
 * In-app notification for a company user. Created automatically by the platform on key events.

 */
export interface Notification {
  id: string;
  company_id: string;
  /** Stable internal type key (e.g. offer_received, outbid, offer_accepted). */
  type: string;
  title_ar: string;
  title_en: string;
  body_ar?: string;
  body_en?: string;
  is_read: boolean;
  /** Entity type the notification links to (e.g. listing, deal) */
  related_entity_type?: string;
  /** UUID of the related entity */
  related_entity_id?: string;
  created_at: Date;
  read_at?: Date;
}